announcement-sender/requirements.txt
belay==0.25.0
pyserial==3.5